server: refactor cluster related logic (#1654)
* refactor cluster related logic
