+++ server/heartbeat_streams.go
@@ -82,7 +82,7 @@ func (s *heartbeatStreams) run() {
 		case msg := <-s.msgCh:
 			storeID := msg.GetTargetPeer().GetStoreId()
 			storeLabel := strconv.FormatUint(storeID, 10)
-			store, err := s.cluster.GetStore(storeID)
+			store, err := s.cluster.TryGetStore(storeID)
 			if err != nil {
 				log.Error("fail to get store",
 					zap.Uint64("region-id", msg.RegionId),
@@ -109,7 +109,7 @@ func (s *heartbeatStreams) run() {
 			}
 		case <-keepAliveTicker.C:
 			for storeID, stream := range s.streams {
-				store, err := s.cluster.GetStore(storeID)
+				store, err := s.cluster.TryGetStore(storeID)
 				if err != nil {
 					log.Error("fail to get store", zap.Uint64("store-id", storeID), zap.Error(err))
 					delete(s.streams, storeID)
