+++ tests/cluster.go
@@ -248,14 +248,14 @@ func (s *TestServer) GetEtcdClient() *clientv3.Client {
 func (s *TestServer) GetStores() []*metapb.Store {
 	s.RLock()
 	defer s.RUnlock()
-	return s.server.GetRaftCluster().GetStores()
+	return s.server.GetRaftCluster().GetMetaStores()
 }
 
 // GetStore returns the store with a given store ID.
 func (s *TestServer) GetStore(storeID uint64) (*core.StoreInfo, error) {
 	s.RLock()
 	defer s.RUnlock()
-	return s.server.GetRaftCluster().GetStore(storeID)
+	return s.server.GetRaftCluster().TryGetStore(storeID)
 }
 
 // GetRaftCluster returns Raft cluster.
@@ -277,7 +277,7 @@ func (s *TestServer) GetRegions() []*core.RegionInfo {
 func (s *TestServer) GetRegionInfoByID(regionID uint64) *core.RegionInfo {
 	s.RLock()
 	defer s.RUnlock()
-	return s.server.GetRaftCluster().GetRegionInfoByID(regionID)
+	return s.server.GetRaftCluster().GetRegion(regionID)
 }
 
 // GetAdjacentRegions returns regions' information that are adjacent with the specific region ID.
